feat : 서비스 인터페이스 -> 클래스로 변경
인터페이스 사용으로 프로젝트 레벨을 복잡해져 class로 변경
