+++ src/main/java/com/grablunchtogether/controller/UserController.java
@@ -1,10 +1,8 @@
 package com.grablunchtogether.controller;
 
-import com.grablunchtogether.common.results.responseResult.ResponseError;
 import com.grablunchtogether.common.results.responseResult.ResponseResult;
 import com.grablunchtogether.common.results.serviceResult.ServiceResult;
 import com.grablunchtogether.dto.clovaOcr.OcrApiDto;
-import com.grablunchtogether.dto.user.UserOcrSignUpInput;
 import com.grablunchtogether.dto.geocode.GeocodeDto;
 import com.grablunchtogether.dto.user.*;
 import com.grablunchtogether.service.externalApi.clovaOcr.OcrApiService;
@@ -15,16 +13,11 @@ import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpHeaders;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.transaction.annotation.Transactional;
-import org.springframework.validation.Errors;
-import org.springframework.validation.FieldError;
 import org.springframework.web.bind.annotation.*;
 
 import javax.validation.Valid;
-import java.util.ArrayList;
-import java.util.List;
 
 @RequiredArgsConstructor
 @RequestMapping("/api/users")
@@ -41,13 +34,7 @@ public class UserController {
     @Transactional
     @ApiOperation(value = "사용자 회원가입", notes = "입력된 정보로 회원가입을 진행합니다.")
     public ResponseEntity<?> userSignUp(
-            @Valid @RequestBody UserSignUpInput userSignUpInput,
-            Errors errors) {
-
-        ResponseEntity<?> responseErrorList = errorValidation(errors);
-        if (responseErrorList != null) {
-            return responseErrorList;
-        }
+            @Valid @RequestBody UserSignUpInput userSignUpInput) {
 
         //고객 좌표 가져오는 외부 api호출
         GeocodeDto userCoordinate = geocodeApiService.getCoordinate(
@@ -61,13 +48,7 @@ public class UserController {
 
     @PostMapping("/login")
     @ApiOperation(value = "사용자 로그인", notes = "입력된 사용자 ID/PW로 로그인을 진행합니다.")
-    public ResponseEntity<?> login(
-            @Valid @RequestBody UserLoginInput userLoginInput, Errors errors) {
-
-        ResponseEntity<?> responseErrorList = errorValidation(errors);
-        if (responseErrorList != null) {
-            return responseErrorList;
-        }
+    public ResponseEntity<?> login(@Valid @RequestBody UserLoginInput userLoginInput) {
 
         ServiceResult result = userService.login(userLoginInput);
 
@@ -78,13 +59,7 @@ public class UserController {
     @ApiOperation(value = "사용자 정보 수정", notes = "입력된 정보로 기존사용자의 정보를 수정합니다.")
     public ResponseEntity<?> editUserInformation(
             @RequestHeader(HttpHeaders.AUTHORIZATION) String token,
-            @Valid @RequestBody UserInformationEditInput userInformationEditInput,
-            Errors errors) {
-
-        ResponseEntity<?> responseErrorList = errorValidation(errors);
-        if (responseErrorList != null) {
-            return responseErrorList;
-        }
+            @Valid @RequestBody UserInformationEditInput userInformationEditInput) {
 
         UserDto userDto = userService.tokenValidation(token);
 
@@ -104,13 +79,7 @@ public class UserController {
     @ApiOperation(value = "사용자 비밀번호 변경", notes = "기존 비밀번호를 입력된 비밀번호로 변경합니다.")
     public ResponseEntity<?> changePassword(
             @RequestHeader(HttpHeaders.AUTHORIZATION) String token,
-            @Valid @RequestBody UserChangePasswordInput userChangePasswordInput,
-            Errors errors) {
-
-        ResponseEntity<?> responseErrorList = errorValidation(errors);
-        if (responseErrorList != null) {
-            return responseErrorList;
-        }
+            @Valid @RequestBody UserChangePasswordInput userChangePasswordInput) {
 
         UserDto userDto = userService.tokenValidation(token);
 
@@ -123,13 +92,7 @@ public class UserController {
     @PostMapping("/signup/ocr")
     @ApiOperation(value = "명함 OCR 회원가입", notes = "명함이미지를 받아 간편회원가입을 진행합니다.")
     public ResponseEntity<?> ocrSignUp(
-            @Valid @RequestBody UserOcrSignUpInput userOcrSignUpInput,
-            Errors errors){
-
-        ResponseEntity<?> responseErrorList = errorValidation(errors);
-        if (responseErrorList != null) {
-            return responseErrorList;
-        }
+            @Valid @RequestBody UserOcrSignUpInput userOcrSignUpInput) {
 
         OcrApiDto ocrData = ocrApiService.getUserInfoFromNameCard(userOcrSignUpInput.getImageName());
 
@@ -149,24 +112,14 @@ public class UserController {
 
         return ResponseResult.result(result);
     }
+
     @PostMapping("/password/reset")
     @ApiOperation(value = "비밀번호 초기화", notes = "비밀번호를 초기화 하고 이메일로 전송합니다.")
     public ResponseEntity<?> ocrSignUp(
-            @RequestBody UserPasswordResetInput userPasswordResetInput){
+            @RequestBody UserPasswordResetInput userPasswordResetInput) {
 
         ServiceResult result =
                 mailSenderService.resetPassword(userPasswordResetInput);
         return ResponseResult.result(result);
-    }
-
-    private ResponseEntity<?> errorValidation(Errors errors) {
-        List<ResponseError> responseErrorList = new ArrayList<>();
-        if (errors.hasErrors()) {
-            errors.getAllErrors().forEach(error -> {
-                responseErrorList.add(ResponseError.of((FieldError) error));
-            });
-            return new ResponseEntity<>(responseErrorList, HttpStatus.BAD_REQUEST);
-        }
-        return null;
     }
 }
