+++ src/main/java/com/grablunchtogether/controller/UserReviewController.java
@@ -1,6 +1,5 @@
 package com.grablunchtogether.controller;
 
-import com.grablunchtogether.common.results.responseResult.ResponseError;
 import com.grablunchtogether.common.results.responseResult.ResponseResult;
 import com.grablunchtogether.common.results.serviceResult.ServiceResult;
 import com.grablunchtogether.dto.user.UserDto;
@@ -11,15 +10,10 @@ import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpHeaders;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.validation.Errors;
-import org.springframework.validation.FieldError;
 import org.springframework.web.bind.annotation.*;
 
 import javax.validation.Valid;
-import java.util.ArrayList;
-import java.util.List;
 
 @RequiredArgsConstructor
 @Api(tags = "User Review API", description = "사용자 리뷰 관련된 API")
@@ -34,13 +28,7 @@ public class UserReviewController {
     public ResponseEntity<?> addReview(
             @PathVariable Long planHistoryId,
             @RequestHeader(HttpHeaders.AUTHORIZATION) String token,
-            @Valid @RequestBody UserReviewInput userReviewInput,
-            Errors errors) {
-
-        ResponseEntity<?> responseErrorList = errorValidation(errors);
-        if (responseErrorList != null) {
-            return responseErrorList;
-        }
+            @Valid @RequestBody UserReviewInput userReviewInput) {
 
         UserDto userDto = userService.tokenValidation(token);
 
@@ -91,16 +79,5 @@ public class UserReviewController {
                 userReviewService.listReviews(targetUserId);
 
         return ResponseResult.result(result);
-    }
-
-    private ResponseEntity<?> errorValidation(Errors errors) {
-        List<ResponseError> responseErrorList = new ArrayList<>();
-        if (errors.hasErrors()) {
-            errors.getAllErrors().forEach(error -> {
-                responseErrorList.add(ResponseError.of((FieldError) error));
-            });
-            return new ResponseEntity<>(responseErrorList, HttpStatus.BAD_REQUEST);
-        }
-        return null;
     }
 }
