+++ src/main/java/com/grablunchtogether/dto/plan/PlanDto.java
@@ -1,6 +1,6 @@
 package com.grablunchtogether.dto.plan;
 
-import com.grablunchtogether.domain.enums.PlanStatus;
+import com.grablunchtogether.enums.PlanStatus;
 import com.grablunchtogether.domain.Plan;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
