+++ src/main/java/com/grablunchtogether/exception/ErrorCode.java
@@ -14,9 +14,9 @@ public enum ErrorCode {
     USER_SIGN_IN_FAIL(HttpStatus.BAD_REQUEST, "아이디 또는 비밀번호가 일치하지않습니다."),
     USER_INFO_NOT_FOUND(HttpStatus.NOT_FOUND, "회원정보를 찾을 수 없습니다."),
     UNDEFINED_EXCEPTION(HttpStatus.BAD_REQUEST, "정의되지 않은 예외가 발생했습니다. 메세지를 참고해주세요."),
-    TOKEN_IS_INVALID(HttpStatus.BAD_REQUEST, ""),
+    TOKEN_IS_INVALID(HttpStatus.BAD_REQUEST, "토큰이 만료되었거나 유효하지 않습니다. 다시 로그인해주세요."),
     PLAN_TIME_NOT_MATCH(HttpStatus.BAD_REQUEST, "약속시간 1시간 이전에만 가능합니다."),
-    USER_REVIEW_ALREADY_EXISTS(HttpStatus.BAD_REQUEST, ""),
+    USER_REVIEW_ALREADY_EXISTS(HttpStatus.BAD_REQUEST, "이미 존재하는 회원입니다."),
     CONTENT_NOT_FOUND(HttpStatus.NOT_FOUND, ""),
     NOT_PERMITTED(HttpStatus.BAD_REQUEST, "접근 권한이 없습니다."),
     CRAWLING_IS_IN_PROGRESS(HttpStatus.BAD_REQUEST, "맛집정보를 업데이트하는 중입니다. 잠시후 다시 시도해주세요."),
