+++ src/main/java/com/grablunchtogether/service/user/UserTokenDetailService.java
@@ -1,0 +1,38 @@
+package com.grablunchtogether.service.user;
+
+import com.grablunchtogether.exception.CustomException;
+import com.grablunchtogether.exception.ErrorCode;
+import com.grablunchtogether.repository.UserRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.userdetails.User;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+import java.util.ArrayList;
+import java.util.List;
+
+@RequiredArgsConstructor
+@Service
+public class UserTokenDetailService implements UserDetailsService {
+
+    private final UserRepository userRepository;
+
+    @Transactional(readOnly = true)
+    @Override
+    public UserDetails loadUserByUsername(String email) throws UsernameNotFoundException {
+        com.grablunchtogether.domain.User userEntity = userRepository.findByUserEmail(email)
+                .orElseThrow(() -> new CustomException(ErrorCode.USER_INFO_NOT_FOUND));
+
+        List<GrantedAuthority> authorities = new ArrayList<>(userEntity.getUserRole().getAuthorities());
+
+        return User.builder()
+                .username(userEntity.getUserEmail())
+                .password(userEntity.getUserPassword())
+                .authorities(authorities)
+                .build();
+    }
+}
